requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.10.0
    # via
    #   feedfinder2
    #   google
    #   newspaper3k
    #   snscrape
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.3
    # via nltk
colorama==0.4.4
    # via
    #   click
    #   tqdm
cssselect==1.1.0
    # via newspaper3k
feedfinder2==0.0.4
    # via newspaper3k
feedparser==6.0.8
    # via newspaper3k
filelock==3.6.0
    # via
    #   snscrape
    #   tldextract
google==3.0.0
    # via -r requirements.in
idna==3.3
    # via
    #   requests
    #   tldextract
inflect==5.4.0
    # via -r requirements.in
jieba3k==0.35.1
    # via newspaper3k
joblib==1.1.0
    # via nltk
lxml==4.8.0
    # via
    #   newspaper3k
    #   snscrape
newspaper3k==0.2.8
    # via -r requirements.in
nltk==3.7
    # via newspaper3k
pillow==9.0.1
    # via newspaper3k
pysocks==1.7.1
    # via requests
python-dateutil==2.8.2
    # via newspaper3k
pytz==2021.3
    # via snscrape
pyyaml==6.0
    # via newspaper3k
regex==2022.1.18
    # via nltk
requests[socks]==2.27.1
    # via
    #   feedfinder2
    #   newspaper3k
    #   requests-file
    #   snscrape
    #   tldextract
requests-file==1.5.1
    # via tldextract
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   feedfinder2
    #   python-dateutil
    #   requests-file
snscrape==0.4.3.20220106
    # via -r requirements.in
soupsieve==2.3.1
    # via beautifulsoup4
tinysegmenter==0.3
    # via newspaper3k
tldextract==3.1.2
    # via newspaper3k
tqdm==4.62.3
    # via nltk
urllib3==1.26.8
    # via requests
